Add "--year all" to run all years in one go.

+++ year2021/src/solver02.rs
@@ -1,4 +1,4 @@
 // TODO
 pub fn solve02(_input: &[String]) -> (i128, i128) {
-    (0 as i128, 0 as i128)
+    (0_i128, 0_i128)
 }
