+++ year2021/src/solver20.rs
@@ -1,4 +1,4 @@
 // TODO
 pub fn solve20(_input: &[String]) -> (i128, i128) {
-    (0 as i128, 0 as i128)
+    (0_i128, 0_i128)
 }
